refactor: modify forms
